Simple modulo optimization
As we are using a base 2 for our RHS modulo operator, we changed that for a bitwise and operator, which is faster.

In other words, if the RHS of the modulo operator (i.e. LHS % RHS) is a base 2, we can write "LHS & (RHS -1)", which will yield the same results but with a faster processing time.